[UPD] It constructs a dictionary, based on the split name for metrics and generalize the code using dictionaries. Generalize PoseClassificationDataset the using **kwargs. Replace the old train.py and pred.py by the code that was on the now-deleted train_debug.py and pred_debug.py files
